### 2025-05-17 #8227 @editorialbot
[PRE REVIEW]: Gene Fetch: A Python tool for high-throughput sequence retrieval from NCBI databases
**Submitting author:** @SchistoDan (<a href="http://orcid.org/0000-0002-5246-0725">Daniel A.J Parsons</a>)
**Repository:** https://github.com/bge-barcoding/gene_fetch
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.9
