### 2025-05-21 d46f6ae
Update gene_fetch.sh
