### 2025-05-23 6ced957
Update paper.bib and paper.md
