+++ paper/paper.bib
@@ -96,7 +96,7 @@
 	number = {3},
 	pages = {725-738},
 	keywords = {aDNA, ancient DNA, eDNA, environmental DNA, python, reference database curation, taxonomy assignment},
-	doi = {https://doi.org/10.1111/1755-0998.13741},
+	doi = {10.1111/1755-0998.13741},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1755-0998.13741},
 	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1755-0998.13741},
 	abstract = {Abstract The measurement of biodiversity is an integral aspect of life science research. With the establishment of second- and third-generation sequencing technologies, an increasing amount of metabarcoding data is being generated as we seek to describe the extent and patterns of biodiversity in multiple contexts. The reliability and accuracy of taxonomically assigning metabarcoding sequencing data have been shown to be critically influenced by the quality and completeness of reference databases. Custom, curated, eukaryotic reference databases, however, are scarce, as are the software programs for generating them. Here, we present crabs (Creating Reference databases for Amplicon-Based Sequencing), a software package to create custom reference databases for metabarcoding studies. crabs includes tools to download sequences from multiple online repositories (i.e., NCBI, BOLD, EMBL, MitoFish), retrieve amplicon regions through in silico PCR analysis and pairwise global alignments, curate the database through multiple filtering parameters (e.g., dereplication, sequence length, sequence quality, unresolved taxonomy, inclusion/exclusion filter), export the reference database in multiple formats for immediate use in taxonomy assignment software, and investigate the reference database through implemented visualizations for diversity, primer efficiency, reference sequence length, database completeness and taxonomic resolution. crabs is a versatile tool for generating curated reference databases of user-specified genetic markers to aid taxonomy assignment from metabarcoding sequencing data. crabs can be installed via docker and is available for download as a conda package and via GitHub (https://github.com/gjeunen/reference\_database\_creator).},
