### 2025-06-12 #8227 @abhishektiwari
Thanks @SchistoDan. Can you action following?

- [ ] Reduce paper Wordcount close to 1000 words (excluding front matter its currently 1174 words).
- [ ] I don't see a community/contribution guidelines. Please add it and mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).
