### 2025-06-19 9e094cb
word-count reduction and addition of contribution section to README
