+++ paper/paper.md
@@ -25,11 +25,11 @@ bibliography: paper.bib
 
 # Summary
 
-`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) sequence databases, namely GenBank [@Benson:2012]. GenBank is one of three mirrored partner databases, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], that form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. As of writing, GenBank comprises >3.6 billion nucleotide sequences from >500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. 
+`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) sequence databases, namely GenBank [@Benson:2012]. GenBank is one of three mirrored partner databases which, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. GenBank currently holds over 3.6 billion nucleotide sequences from over 500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. 
 
 By streamlining the often tedious and error-prone task of sequence acquisition at scale, `Gene Fetch` addresses a critical need in biological research. It is capable of retrieving both protein and nucleotide sequences for a user-specified genetic marker from across the tree of life, including protein-coding genes (such as cytochrome oxidase subunits, NADH dehydrogenases, RuBisCO and matK), ribosomal RNA genes (like 16S, 18S, and 28S) and the internal transcribed spacer (ITS) regions. The tool accepts both NCBI taxonomic identifiers (taxids) and hierarchical taxonomic information as input. 
 
-It can systematically traverse taxonomic hierarchies when target sequences are unavailable at the initially specified taxonomic rank (e.g., species → genus → family, etc), ensuring retrieval of the taxonomically closest available reference sequence. This functionality is particularly valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data, and serves to “fetch” the closest available sequence for each of the focal taxa. The integrated ‘batch’ mode processes multiple input taxa and retrieves the ‘best’ sequence per taxa, while the ‘single’ mode enables exhaustive searching and retrieval of sequence targets for a specified taxon. Collectively, these two modes facilitate more efficient retrieval of sequence data for genomic and phylogenetic studies across diverse taxa.
+It can systematically traverse taxonomic hierarchies when target sequences are unavailable at the initially specified taxonomic rank (e.g., species → genus → family, etc), ensuring retrieval of the taxonomically closest available reference sequence. This is especially valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data, facilitating retrieval of the taxonomically closest available sequence. The integrated ‘batch’ mode processes multiple input taxa and retrieves the ‘best’ sequence per taxa, while the ‘single’ mode exhaustively searches for target sequences for a specified taxon. Collectively, these modes enable efficient retrieval of sequence data for genomic and phylogenetic studies across diverse taxa.
 
 
 
@@ -41,24 +41,24 @@ Comparative genetic analyses within or between taxonomic groups often requires r
 (2) Variable sequence representation across taxonomic groups on databases.
 (3) Time-intensive manual retrieval processes that do not scale efficiently.
 
[…]
