### 2025-06-19 #8227 @SchistoDan
@abhishektiwari Sincerest apologies for the delay in addressing your previous issue.

I have now updated the README with a brief contributions and guidlines section following the CONTRIBUTING.md example you shared. I've also reduced the wordcount closer to 1,000 words. These updates have been pushed to the repo.

Many thanks,
Dan
