### 2025-06-24 #8227 @rvosa
I'd be happy to!
