### 2025-06-27 #8456 @Kamouyiaraki
@abhishektiwari Regarding my COI: I work at the same institution with both authors, and I am currently working with Benjamin Price on a genome sequencing project. I have also offered advice to the lead author on this work. I have no overlap with the content of this paper in my own work and feel that I can provide an honest review, however I understand that I may be too close to this. Apologies for the inconvenience.
