### 2025-06-28 #8456 @abhishektiwari
👋👋 @johnvusich @skanwal @jeskowagner, would any of you be willing to review this submission for JOSS?  The submission under consideration is "Gene Fetch: A Python tool for high-throughput sequence retrieval from NCBI databases".

Based on your experience, we think you might be able to provide a great review of this submission.

Thanks
Abhi
