### 2025-06-29 #8456 @rvosa
> Thanks [@Kamouyiaraki](https://github.com/Kamouyiaraki) [@rvosa](https://github.com/rvosa) for being transparent about the COI. Unfortunately, after discussion with EiC, I have decided to bring new reviewers for this submissions. Thanks again for your time, I will look forward to working with you on other JOSS submissions.

I understand completely. This is how the system should work and how we uphold integrity in the process. I look forward to the review process being completed satisfactorily because, in my **biased** (🤣) opinion, this is good work.
