### 2025-07-01 #8456 @abhishektiwari
@hrhotz  thanks for agreeing to review this submission. Your review will be checklist based.

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know.
