### 2025-07-01 #8456 @johnvusich
### Minor Text Edits Suggested

---

**1. Original:**

> User’s can also specify additional markers and synonyms.

**Suggested:**

> Users can also specify additional markers and synonyms.

**Reason:** Incorrect use of possessive apostrophe; should be the plural form "Users".

---

**2. Original:**

> ...capable of retrieving both protein and nucleotide sequences for a user-specified genetic marker from across the tree of life, including protein-coding genes (such as cytochrome oxidase subunits, NADH dehydrogenases, RuBisCO and matK), ribosomal RNA genes (like 16S, 18S, and 28S) and the internal transcribed spacer (ITS) regions.

**Suggested:**

> ...including protein-coding genes (such as cytochrome oxidase subunits, NADH dehydrogenases, RuBisCO, and matK), ribosomal RNA genes (like 16S, 18S, and 28S), and the internal transcribed spacer (ITS) regions.

**Reason:** Added serial commas for clarity and grammatical consistency.

---

**3. Original:**

> Optimised for 15 common genetic markers, , including “barcoding” genes...

**Suggested:**

> Optimised for 15 common genetic markers, including “barcoding” genes...

**Reason:** Duplicate comma after "markers".

---

**4. Original:**

> Gene Fetch can process hundereds-thousands of samples...

**Suggested:**

> Gene Fetch can process hundreds to thousands of samples...

**Reason:** Typo ("hundereds") and better phrasing with "to".

---

**5. Original:**

> ...retrieval of the taxonomically closest available reference sequence. This is especially valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data, facilitating retrieval of the taxonomically closest available sequence.

**Suggested:**

> ...retrieval of the taxonomically closest available reference sequence. This is especially valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data.

**Reason:** Repetition of “retrieval of the taxonomically closest available sequence.”

---

**6. Original:**

> Gene Fetch offers several features: - Traverses taxonomic hierarchies...

**Suggested:**

> Gene Fetch offers several features:
>
> - Traverses taxonomic hierarchies...

**Reason:** Cleaned up formatting for the list (moved dash to next line).

---

**7. Original:**

> Includes fallback mechanisms for handling inconsistently annotated entries.

**Suggested:**

> Includes fallback mechanisms to handle inconsistently annotated entries.

**Reason:** Slightly smoother and more active construction.

---

**8. Original:**

> User's are encouraged to report bugs and suggest enhancements through the GitHub issue tracker.

**Suggested:**

> Users are encouraged to report bugs and suggest enhancements through the GitHub issue tracker.

**Reason:** Incorrect use of possessive apostrophe; should be plural "Users".

---

### Testing and reproducibility feedback

| Category                          | Outcome / Feedback |
|----------------------------------|---------------------|
| **Installation via pip**         | Successful in fresh Conda environment using `pip install gene-fetch` |
| **Command-line functionality**   | `gene-fetch` CLI works with real input and produces expected FASTA and GenBank outputs |
| **Test suite**                   | `pytest` fails due to broken mocks expecting a deprecated function `setup_argument_parser` |
| **Dependencies**                 | No `requirements.txt` or `environment.yml`; test dependencies undocumented |
| **GitHub SLURM script**          | `gene_fetch.sh` script does not work out-of-the-box — `gene-fetch` invocation is incomplete (missing `--email`, `--api-key`, and flags not positioned correctly) |
| **Conda guidance in README**     | Recommends Conda but does not provide full list of dependencies or tested environment file |
| **Continuous Integration (CI)**  | No CI workflow included to validate installation or testing automatically |

**Suggestions:**
- Update or remove broken unit tests that rely on deprecated functions
- Provide a minimal working `environment.yml` or `requirements-dev.txt` for reproducibility
- Fix or revise the SLURM example script (`gene_fetch.sh`) to include the full `gene-fetch` command with appropriate flags and arguments
- Consider adding a GitHub Actions workflow to test CLI execution and key functions
