### 2025-07-01 a6fe755
Update paper.md
