+++ README.md
@@ -42,7 +42,9 @@ Gene Fetch enables high-throughput retreival of sequence data from NCBI database
 
 
 ## Installation
-- Due to the risk of dependency conflicts, it's recommended to install Gene Fetch in a Conda environment. First Conda needs to be installed, which can be done from [here](https://www.anaconda.com/docs/getting-started/miniconda/install). Once installed:
+- Due to the risk of dependency conflicts, it's recommended to install Gene Fetch in a Conda environment.
+- First Conda needs to be installed, which can be done from [here](https://www.anaconda.com/docs/getting-started/miniconda/install).
+- Once installed:
 ```bash
 # Create new environment
 conda create -n gene-fetch
@@ -51,13 +53,13 @@ conda create -n gene-fetch
 conda activate gene-fetch
 ```
 
-- Gene Fetch can then be installed from [PyPI](https://pypi.org/project/gene-fetch/#description), or [Bioconda](https://anaconda.org/bioconda/gene-fetch):
+- Gene Fetch can then be installed from [Bioconda](https://anaconda.org/bioconda/gene-fetch), or [PyPI](https://pypi.org/project/gene-fetch/#description):
 ```bash
-# Install via pip
-pip install gene-fetch
-
-# OR, Install via bioconda
+# Install via bioconda
 conda install bioconda::gene-fetch
+
+# Or, install via pip
+pip install gene-fetch
 
 # Verify installation
 gene-fetch --help
@@ -69,7 +71,7 @@ gene-fetch --help
 git clone https://github.com/bge-barcoding/gene_fetch.git
 cd gene_fetch
 
-#Install dependencies (likely in a conda env)
+#Install dependencies (listed in environment.yaml)
 - biopython=1.85
 - ratelimit=2.2.1
 - gene-fetch=1.0.11
