### 2025-07-02 bd14248
Update gene_fetch.sh
