### 2025-07-02 #8456 @abhishektiwari
@ManavalanG  thanks for agreeing to review this submission. Your review will be checklist based.

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know.
