### 2025-07-03 687c1e2
Create test.yml
Implementing GitHub Actions workflow to test CLI execution and key functions
