+++ src/gene_fetch/processors.py
@@ -1,1869 +1,496 @@
-# src/gene_fetch/sequence_processor.py
+# src/gene_fetch/processors.py
 """
-SequenceProcessor class for Gene Fetch.
-Handles the processing and validation of fetched sequences.
+Processing functions for Gene Fetch.
+Contains functions for batch and single-taxid processing.
 """
 
-import re
+import csv
 import logging
-from time import sleep
+import sys
+import traceback
+from pathlib import Path
 from random import uniform
-from typing import Optional, Tuple, List, Dict, Any
+from time import sleep
+from typing import Optional, Dict
 
-from Bio import Entrez, SeqIO
-from Bio.Seq import Seq
+from Bio import SeqIO
 from Bio.SeqRecord import SeqRecord
 
-from .core import logger, Config
-from .entrez_handler import EntrezHandler, check_ncbi_status
+from .core import logger, make_out_dir, log_progress, get_process_id_column
+from .entrez_handler import EntrezHandler
+from .sequence_processor import SequenceProcessor
+from .output_manager import OutputManager, save_genbank_file
 
 
 # =============================================================================
-# Processing and validation of fetched entries and sequences
+# Processing functions
 # =============================================================================
-class SequenceProcessor:
-    def __init__(self, config: Config, entrez: EntrezHandler):
-        self.config = config
-        self.entrez = entrez
+# Process input per sample, retrieving and storing fetched sequences
+def process_sample(
+    process_id: str,
+    taxid: str,
+    sequence_type: str,
+    processor: SequenceProcessor,
+    output_manager: OutputManager,
+    gene_name: str,
+    save_genbank: bool = False,
+) -> None:
+    try:
+        # Define output paths
+        protein_path = output_manager.output_dir / f"{process_id}.fasta"
+        nucleotide_path = output_manager.nucleotide_dir / f"{process_id}_dna.fasta"
 
-    # Handle 'contig' lines in feature table
-    def parse_contig_line(self, contig_line: str) -> Optional[Tuple[str, int, int]]:
-        try:
-            # Remove 'join(' and ')' if present
-            cleaned = contig_line.strip().replace("join(", "").replace(")", "")
+        # Define GenBank paths if needed
+        protein_gb_path = output_manager.genbank_dir / f"{process_id}.gb"
+        nucleotide_gb_path = output_manager.genbank_dir / f"{process_id}_dna.gb"
 
-            # Parse the contig reference. Format is typically: WVEN01000006.2:1..16118
-            if ":" in cleaned:
-                contig_id, coords = cleaned.split(":")
-                if ".." in coords:
-                    start, end = coords.split("..")
-                    return contig_id, int(start), int(end)
-        except Exception as e:
-            logger.error(f"Error parsing CONTIG line '{contig_line}': {e}")
-        return None
-
-    # Handle WGS records that might not contain sequence data, fetching associated sequence if listed in 'contig' line
-    def fetch_wgs_sequence(self, record: SeqRecord) -> Optional[SeqRecord]:
-        try:
-            # Check for CONTIG line in annotations
-            contig_line = record.annotations.get("contig", None)
-            if not contig_line:
-                logger.warning(f"No CONTIG line found in WGS record {record.id}")
-                return None
-
-            # Parse the CONTIG line
-            contig_info = self.parse_contig_line(contig_line)
-            if not contig_info:
-                logger.error(f"Could not parse CONTIG line: {contig_line}")
-                return None
-
-            contig_id, start, end = contig_info
-            logger.info(
-                f"Found WGS contig reference: {contig_id} positions {start}..{end}"
-            )
-
-            # Fetch the actual contig sequence
-            try:
-                handle = self.entrez.fetch(
-                    db="nucleotide",
[…]
