### 2025-07-03 bge-barcoding/gene_fetch#6 @hrhotz
comments about the JOSS article
While I am looking into the functionality of you impressive tool, I would like to make a few suggestions wrt the article text - https://github.com/openjournals/joss-reviews/issues/8456

 - I think the title is misleading (i.e.: your tool does much more than the title suggests). It would good to add something like: "across the tree of life" or "traverse taxonomic hierarchies". Maybe: "Gene Fetch: A Python tool for sequence retrieval from GenBank across the tree of life"

 - line 8: even though GenBank suggest to reference the old paper from the 2013 NAR database issue, I would add the current one as well: https://doi.org/10.1093/nar/gkae1114

 - line 52: supports a “batch” and “single” query modes  is repeated on line 66: Supports “batch” mode and “single” mode

 - similar, line 53: length filtering is repeated on line 68

 - I would merge the two section (lines 52 to 60 and lines 61 to 72), remove all the duplications and add a simple example instead

 - line 88/89: The GitHub and package repositories also hold comprehensive... This is obvious, if the article has been accepted by JOSS
