+++ src/gene_fetch/__init__.py
@@ -11,6 +11,3 @@ from .core import Config
 from .entrez_handler import EntrezHandler
 from .sequence_processor import SequenceProcessor
 from .output_manager import OutputManager, save_genbank_file
-
-# Expose main entry point for usage
-from .main import main
