### 2025-07-03 ad5b45a
Update core.py
