+++ .github/workflows/test.yml
@@ -34,21 +34,38 @@ jobs:
       run: |
         poetry run gene-fetch --help
         echo "✓ CLI help command works"
-
+        
     - name: Test CLI parameter validation (expected failures)
       run: |
         # Test missing required parameters
         poetry run gene-fetch -g cox1 --type protein -o ./test_output && exit 1 || echo "✓ CLI fails gracefully without email/api-key"
         
-        # Test invalid gene parameter
-        poetry run gene-fetch -e "test@example.com" -k "fake_key" -g invalid_gene --type protein -i tests/data/test_samples.csv -o ./test_output && exit 1 || echo "✓ CLI handles invalid gene names"
+        # Test invalid email format
+        # poetry run gene-fetch -e "invalid-email" -k "valid_test_key_12345" -g cox1 --type protein -i tests/data/test_samples.csv -o ./test_output && exit 1 || echo "✓ CLI handles invalid email format"
         
-        # Test invalid type parameter
-        poetry run gene-fetch -e "test@example.com" -k "fake_key" -g cox1 --type invalid_type -i tests/data/test_samples.csv -o ./test_output && exit 1 || echo "✓ CLI handles invalid sequence types"
-        
-        # Test missing input file
-        poetry run gene-fetch -e "test@example.com" -k "fake_key" -g cox1 --type protein -i nonexistent.csv -o ./test_output && exit 1 || echo "✓ CLI handles missing input files"
+        # Test invalid API key
+        # poetry run gene-fetch -e "test@example.com" -k "fake_key" -g cox1 --type protein -i tests/data/test_samples.csv -o ./test_output && exit 1 || echo "✓ CLI handles invalid API key"
 
+    - name: Test CLI with real credentials (file validation)
+      env:
+        NCBI_EMAIL: ${{ secrets.NCBI_EMAIL }}
+        NCBI_API_KEY: ${{ secrets.NCBI_API_KEY }}
+      run: |
+        # Only run if secrets are available
+        if [ -n "$NCBI_EMAIL" ] && [ -n "$NCBI_API_KEY" ]; then
+          echo "Testing with real NCBI credentials..."
+          
+          # Test missing input file (should fail for file reasons, not credential reasons)
+          poetry run gene-fetch -e "$NCBI_EMAIL" -k "$NCBI_API_KEY" -g cox1 --type protein -i nonexistent.csv -o ./test_output && exit 1 || echo "✓ CLI handles missing input files"
+          
+          # Test invalid gene with real credentials (should proceed past validation but fail later)
+          timeout 20s poetry run gene-fetch -e "$NCBI_EMAIL" -k "$NCBI_API_KEY" -g invalid_gene_xyz123 --type protein -i tests/data/test_samples.csv -o ./test_output || echo "✓ CLI processes invalid gene names (timed out or failed as expected)"
+          
+        else
+          echo "⚠️  NCBI credentials not available - skipping credential tests"
+          echo "   (This is normal for external contributors)"
+        fi
+        
     - name: Test CLI with test data (parameter parsing)
       run: |
         # Test that CLI accepts valid parameters and creates output directory
