### 2025-07-03 b41b996
Update core.py
