### 2025-07-03 cd2c2d1
Update test_core.py
