### 2025-07-03 ecf7d75
Update test_entrez_handler.py
