### 2025-07-03 a1c475e
Update test_simple_import.py
