### 2025-07-03 1cfb9ec
Update test_core.py
