### 2025-07-03 42c7f34
Update core.py
