+++ src/gene_fetch/core.py
@@ -114,12 +114,16 @@ class Config:
         # Allow specific test API keys for unit testing
         test_api_keys = ["test_api_key_1234567890", "valid_test_key_12345"]
     
-        # Reject obviously fake keys, but allow test keys even if they're short
-        if api_key in ["fake_key", "fake_api_key"]:
+        # If it's a test key, allow it regardless of other checks
+        if api_key in test_api_keys:
+            return
+    
+        # Reject obviously fake keys
+        if api_key in ["fake_key", "fake_api_key", "123456789", "test", "test_key"]:
             raise ValueError("Invalid API key. Please provide a valid NCBI API key.")
     
-        # For non-test keys, enforce minimum length
-        if api_key not in test_api_keys and len(api_key) < 10:
+        # For all other keys (real keys), enforce minimum length
+        if len(api_key) < 10:
             raise ValueError("Invalid API key. Please provide a valid NCBI API key.")
 
     def __init__(self, email, api_key):
