### 2025-07-03 b0a269b
Update test_core.py
