### 2025-07-03 73aaeab
Update core.py
