+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
         poetry run gene-fetch --help
         echo "✓ CLI help command works"
         
-    - name: Test CLI parameter validation (expected failures)
+    - name: Test CLI properly rejects invalid credentials (expected failure)
       run: |
         # Test missing required parameters
         poetry run gene-fetch -g cox1 --type protein -o ./test_output && exit 1 || echo "✓ CLI fails gracefully without email/api-key"
@@ -46,7 +46,7 @@ jobs:
         # Test invalid API key
         # poetry run gene-fetch -e "test@example.com" -k "fake_key" -g cox1 --type protein -i tests/data/test_samples.csv -o ./test_output && exit 1 || echo "✓ CLI handles invalid API key"
 
-    - name: Test CLI with real credentials (file validation)
+    - name: Test CLI accepts valid inputs and rejects invalid ones
       env:
         NCBI_EMAIL: ${{ secrets.NCBI_EMAIL }}
         NCBI_API_KEY: ${{ secrets.NCBI_API_KEY }}
@@ -58,79 +58,60 @@ jobs:
           # Test missing input file (should fail for file reasons, not credential reasons)
           poetry run gene-fetch -e "$NCBI_EMAIL" -k "$NCBI_API_KEY" -g cox1 --type protein -i nonexistent.csv -o ./test_output && exit 1 || echo "✓ CLI handles missing input files"
           
-          echo "✓ All credential and file validation tests passed"
+          echo "✓ Credential and file validation tests passed"
           
         else
           echo "⚠️  NCBI credentials not available - skipping real credential tests"
           echo "   (This is normal for external contributors)"
         fi
         
-    - name: Test CLI with test data (parameter parsing)
+    - name: Test CLI with test data (parameter parsing and execution)
+      env:
+        NCBI_EMAIL: ${{ secrets.NCBI_EMAIL }}
+        NCBI_API_KEY: ${{ secrets.NCBI_API_KEY }}
       run: |
-        # Test that CLI accepts valid parameters and creates output directory
-        # This will likely fail at NCBI API stage, but tests parameter validation
-        mkdir -p cli_test_output
-        
-        # Test with your actual test data
-        timeout 30s poetry run gene-fetch \
-          -e "test@example.com" \
-          -k "fake_api_key" \
-          -g cox1 \
-          --type protein \
-          -i tests/data/test_samples.csv \
-          -o cli_test_output \
-          --protein-size 100 || echo "✓ CLI parameter parsing works (expected API failure)"
-        
-        # Test single mode with valid taxid from your test data
-        timeout 30s poetry run gene-fetch \
-          -e "test@example.com" \
-          -k "fake_api_key" \
-          -g cox1 \
-          --type both \
-          -s 210239 \
-          -o cli_test_output \
-          --max-sequences 2 \
-          --genbank || echo "✓ CLI single mode parameter parsing works (expected API failure)"
-        
-        # Verify output directory structure is created
-        [ -d "cli_test_output" ] && echo "✓ Output directory created" || echo "✗ Output directory not created"
-
-    - name: Verify CLI creates proper file structure
-      run: |
-        # Test that CLI creates expected directory structure
-        poetry run python -c "
-        import os
-        import sys
-        sys.path.insert(0, 'src')
-        
-        # Test the output directory creation logic
-        test_dirs = ['structure_test_protein', 'structure_test_both', 'structure_test_nucleotide']
-        
-        for test_dir in test_dirs:
-            if os.path.exists(test_dir):
-                import shutil
-                shutil.rmtree(test_dir)
-        
-        # This tests internal functions that CLI uses
-        try:
-            from gene_fetch.main import setup_output_directory
-            
[…]
