### 2025-07-03 349d60d
Update core.py
