### 2025-07-03 bge-barcoding/gene_fetch#7 @hrhotz
comments about installation and testing
looking into installation/testing - https://github.com/openjournals/joss-reviews/issues/8456

 - straight forward and successful in fresh Conda environment using conda install bioconda::gene-fetch

 - given the new license restrictions of Anaconda, I would not point to https://www.anaconda.com/docs/getting-started/miniconda/install but rather to another installer, like: https://github.com/conda-forge/miniforge

 - (this could well be a local issue due to our storage system....) when running the test ('pytest') I got in addition to the warning one error.

                #   save a call to os.lstat() when walking subdirectories.
                func, dirfd, path, orig_entry = stack.pop()
                name = path if orig_entry is None else orig_entry.name
                try:
                    if func is os.close:
                        os.close(dirfd)
                        return
                    if func is os.rmdir:
        >               os.rmdir(name, dir_fd=dirfd)
        E               OSError: [Errno 39] Directory not empty: '/scratch/***/tmph6lq_vxh'
