+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## v1.0.12
+- Added credential validation to reject invalid/fake NCBI credentials (email & api key)
+- Implemented automated GitHub Action testing
+- Enhanced test coverage for credential validation
+- Improved CLI error handling and user feedback
