### 2025-07-03 #8456 @SchistoDan
Thanks @johnvusich!

I have addressed the minor text edits you highlighted, added an `environment.yaml` of dependencies and refer to it in the README.md, corrected `gene-fetch.sh` with the proper arguments, added a GitHub Action workflow to test core functionality (all pass), and improved coverage of the testing (e.g. I included additional tests to validate user-specified NCBI credentials).

I have managed to reproduce and overcome the `setup_argument_parser` error, and `pyest` now runs to completion. I've also added a CHANGELOG.md to track all version changes going forward.

I have updated the PyPi package to version 1.0.12 to reflect these improvements, and are currently awaiting reviewer approval for 1.0.12 of the bioconda package.
