### 2025-07-03 bge-barcoding/gene_fetch#6 @hrhotz
...
 - explain the difference between "single" and "batch" mode (especially, since "single" mode can return several sequences as well)
