### 2025-07-04 bge-barcoding/gene_fetch#6 @hrhotz
...

 - is it really necessary to give examples of protein coding genes and ribosomal RNA genes?
