### 2025-07-04 bge-barcoding/gene_fetch#8 @hrhotz
general comments about the usage of gene-fetch
as part of reviewing https://github.com/openjournals/joss-reviews/issues/8456

 - the error message "gene-fetch: error: the following arguments are required: -g/--gene, -o/--out, --type, -e/--email, -k/--api-key" is misleading since either '-i', '-i2' or '-s' is required as well

 - why is there no single character argument for type, protein-size, nucleotide-size, max-sequences? At least the option to use "-t" instead of "--type" would make sense.
