### 2025-07-04 bge-barcoding/gene_fetch#9 @hrhotz
questions about the first example
as part of reviewing https://github.com/openjournals/joss-reviews/issues/8456

 - I can't find the file described in the README: "samples.csv"

 - when using "--type both" and "--genbank" , you get the protein sequences in "output_dir/" and the nucleotide sequences in a subfolder "output_dir/nucleotide/", while the subfolder "output_dir/genbank/" contains both, plus two empty folders "output_dir/genbank/nucleotide/" and "output_dir/genbank/genbank/" - this is confusing.

 - what is the difference between "output_dir/failed_searches.csv" "output_dir/genbank/failed_searches.csv"
