### 2025-07-04 bge-barcoding/gene_fetch#10 @hrhotz
questions about the second example
as part of reviewing https://github.com/openjournals/joss-reviews/issues/8456

I can't find the file described in the README: "samples_taxonomy.csv" or "taxonomy.csv" (as it is called in the example command)

Hence, my self written file looks like:

```
(gene-fetch) [*]$ cat taxonomy.csv
ID,phylum,class,order,family,genus,species
sample-1,Arthropoda,Insecta,Diptera,Acroceridae,Astomella,
sample-2,Arthropoda,Insecta,Hemiptera,Cicadellidae,Psammotettix,Psammotettix sabulicola
sample-3,Arthropoda,Insecta,Trichoptera,Limnephilidae,Dicosmoecus,Dicosmoecus palatus
(gene-fetch) [*]$
```

and when I execute the example command, I get

 - "Missing required genus or species for sample-1"     - the genus is provided, isn't-it?
 - "No valid sequences found for taxID 30147"
 - "No valid sequences found for taxID 1875331"
