### 2025-07-04 bge-barcoding/gene_fetch#11 @hrhotz
skipping existing sequences
as part of reviewing https://github.com/openjournals/joss-reviews/issues/8456

I use the following csv file:

```
(gene-fetch) [*]$ cat test1.csv
ID,taxid
SAMPLE-1,9606
(gene-fetch) [*]$
```

and run the following command:

`gene-fetch -g cbx1 -o output_dir --type nucleotide -e * -k * -i test1.csv
`

now I run the same command but with test2.csv

```
(gene-fetch) [*]$ cat test2.csv
ID,taxid
SAMPLE-1,10090
SAMPLE-2,9606
(gene-fetch) [*]$
```

and the tool returns

```
[INFO] ====== Processing sample 1/2: SAMPLE-1 (taxID: 10090) ======
[INFO] Sequence file(s) already exist for SAMPLE-1. Skipping.
```

and eventually, I end up with the human sequence twice.

```
(gene-fetch) [*]$ diff output_dir/nucleotide/SAMPLE-1_dna.fasta output_dir/nucleotide/SAMPLE-2_dna.fasta
1c1
< >SAMPLE-1
---
> >SAMPLE-2
(gene-fetch) [*]$
```

It would make more sense to test on the taxid (and include it in the fasta header line) rather than on the ID, IMHO.
