### 2025-07-04 bge-barcoding/gene_fetch#12 @hrhotz
Supported targets
What is the advantage of "Supported targets" hard coded in your script? The tool works with "-g 'chromobox 1'" as well as with '-g CBX1' (even the alias 'CBX' works) despite these name variations are not listed.

 - does it make sense to have the "Supported targets" hardcoded in the script? Having an extra file would be easier to handle and is probably less error prone that the requirement to change code in the 'core'.

 - reading the manuscript for the JOSS article (https://github.com/openjournals/joss-reviews/issues/8456), your target audience has no "significant scripting expertise" and you don't want to "burden the user with navigating database-specific syntax and structure". Identifying those "Supported targets" and assign the search fields like "[Gene]","[Title]","[rRNA]" does require advanced knowledge on NCBI search term construction.
