### 2025-07-04 c75382d
Update environment.yaml
