### 2025-07-04 d41a811
paper.bib and paper.md updates (issue 6)
