+++ paper/paper.bib
@@ -14,6 +14,22 @@
     eprint = {https://academic.oup.com/nar/article-pdf/41/D1/D36/3680750/gks1195.pdf},
 }
 
+@article{Sayers:2025,
+    author = {Sayers, Eric W and Cavanaugh, Mark and Frisse, Linda and Pruitt, Kim D and Schneider, Valerie A and Underwood, Beverly A and Yankie, Linda and Karsch-Mizrachi, Ilene},
+    title = {GenBank 2025 update},
+    journal = {Nucleic Acids Research},
+    volume = {53},
+    number = {D1},
+    pages = {D56-D61},
+    year = {2024},
+    month = {11},
+    abstract = {GenBank® (https://www.ncbi.nlm.nih.gov/genbank/) is a comprehensive, public data repository that contains 34 trillion base pairs from over 4.7 billion nucleotide sequences for 581 000 formally described species. Daily data exchange with the European Nucleotide Archive and the DNA Data Bank of Japan ensures worldwide coverage. We summarize the content of the database in 2025 and recent updates such as accelerated processing of influenza sequences and the ability to upload feature tables to Submission Portal for messenger RNA sequences. We provide an overview of the web, application programming and command-line interfaces that allow users to access GenBank data. We also discuss the importance of creating BioProject and BioSample records during submissions, particularly for viruses and metagenomes. Finally, we summarize educational materials and recent community outreach efforts.},
+    issn = {1362-4962},
+    doi = {10.1093/nar/gkae1114},
+    url = {https://doi.org/10.1093/nar/gkae1114},
+    eprint = {https://academic.oup.com/nar/article-pdf/53/D1/D56/60743813/gkae1114.pdf},
+}
+
 @article{OCathail2024,
     author = {O’Cathail, Colman and Ahamed, Alisha and Burgin, Josephine and Cummins, Carla and Devaraj, Rajkumar and Gueye, Khadim and Gupta, Dipayan and Gupta, Vikas and Haseeb, Muhammad and Ihsan, Maira and Ivanov, Eugene and Jayathilaka, Suran and Kadhirvelu, Vishnukumar and Kumar, Manish and Lathi, Ankur and Leinonen, Rasko and McKinnon, Jasmine and Meszaros, Lili and Pauperio, Joana and Pesant, Stephane and Rahman, Nadim and Rinck, Gabriele and Selvakumar, Sandeep and Suman, Swati and Sunthornyotin, Yanisa and Ventouratou, Marianna and Waheed, Zahra and Woollard, Peter and Yuan, David and Zyoud, Ahmad and Burdett, Tony and Cochrane, Guy},
     title = {The European Nucleotide Archive in 2024},
