+++ paper/paper.md
@@ -1,5 +1,5 @@
----
-title: 'Gene Fetch: A Python tool for high-throughput sequence retrieval from NCBI databases'
+---	
+title: 'Gene Fetch: A Python tool for sequence retrieval from GenBank across the tree of life'
 tags:
   - Python
   - Bioinformatics
@@ -25,41 +25,27 @@ bibliography: paper.bib
 
 # Summary
 
-`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) sequence databases, namely GenBank [@Benson:2012]. GenBank is one of three mirrored partner databases which, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. GenBank currently holds over 3.6 billion nucleotide sequences from over 500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. 
+`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) sequence databases, namely GenBank [@Benson:2012; @Sayers:2024]]. GenBank is one of three mirrored partner databases which, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. GenBank currently holds over 3.6 billion nucleotide sequences from over 500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. 
 
 By streamlining the often tedious and error-prone task of sequence acquisition at scale, `Gene Fetch` addresses a critical need in biological research. It is capable of retrieving both protein and nucleotide sequences for a user-specified genetic marker from across the tree of life, including protein-coding genes (such as cytochrome oxidase subunits, NADH dehydrogenases, RuBisCO, and matK), ribosomal RNA genes (like 16S, 18S, and 28S) and the internal transcribed spacer (ITS) regions. The tool accepts both NCBI taxonomic identifiers (taxids) and hierarchical taxonomic information as input. 
 
-It can systematically traverse taxonomic hierarchies when target sequences are unavailable at the initially specified taxonomic rank (e.g., species → genus → family, etc). This is especially valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data, facilitating retrieval of the taxonomically closest available sequence. The integrated ‘batch’ mode processes multiple input taxa and retrieves the ‘best’ sequence per taxa, while the ‘single’ mode exhaustively searches for target sequences for a specified taxon. Collectively, these modes enable efficient retrieval of sequence data for genomic and phylogenetic studies across diverse taxa.
+It can systematically traverse taxonomic hierarchies when target sequences are unavailable at the initially specified taxonomic rank (e.g., species → genus → family, etc), documenting the matched taxonomic rank. This is especially valuable for researchers working with non-model organisms or taxonomic groups with sparse genetic data, facilitating retrieval of the taxonomically closest available sequence. The integrated 'batch' mode processes multiple input taxa and retrieve the single ‘best’ sequence per taxa, whilst 'single' mode exhaustively searches for all target sequences for a specified taxon. Collectively, these modes enable efficient retrieval of sequence data for genomic and phylogenetic studies across diverse taxa.
 
 
 
 
 # Statement of need
 
-Comparative genetic analyses within or between taxonomic groups often requires researchers to gather large numbers of gene or protein sequences from public repositories. Streamlined access to these invaluable databases is essential for comprehensive sequence data analysis, however, this task presents several significant challenges: 
[…]
