### 2025-07-04 3356990
paper.bib and paper.md updates (Issue 6)
