### 2025-07-04 a36b732
Update paper.bib
