### 2025-07-04 bge-barcoding/gene_fetch#6 @SchistoDan
Not at all - It's always good to be sanity checked on these things. I appreciate the thorough review.

Oops! Thanks for spotting and correcting that - I'll merge the correction.
