### 2025-07-04 72c4efa
usage updates (Issue 8)
