### 2025-07-07 89069a4
update in response to issue #9 and #10
