+++ src/gene_fetch/__init__.py
@@ -5,9 +5,9 @@ This package fetches sequence data from NCBI databases
 using sample taxonomic information.
 """
 
-__version__ = "1.0.11"
+__version__ = "1.0.12"
 
 from .core import Config
 from .entrez_handler import EntrezHandler
 from .sequence_processor import SequenceProcessor
-from .output_manager import OutputManager, save_genbank_file
+from .output_manager import OutputManager
