+++ src/gene_fetch/processors.py
@@ -19,7 +19,7 @@ from Bio.SeqRecord import SeqRecord
 from .core import logger, make_out_dir, log_progress, get_process_id_column
 from .entrez_handler import EntrezHandler
 from .sequence_processor import SequenceProcessor
-from .output_manager import OutputManager, save_genbank_file
+from .output_manager import OutputManager
 
 
 # =============================================================================
@@ -37,12 +37,14 @@ def process_sample(
 ) -> None:
     try:
         # Define output paths
-        protein_path = output_manager.output_dir / f"{process_id}.fasta"
-        nucleotide_path = output_manager.nucleotide_dir / f"{process_id}_dna.fasta"
+        protein_path = output_manager.protein_dir / f"{process_id}.fasta"
+        nucleotide_path = output_manager.nucleotide_dir / f"{process_id}.fasta"
 
         # Define GenBank paths if needed
-        protein_gb_path = output_manager.genbank_dir / f"{process_id}.gb"
-        nucleotide_gb_path = output_manager.genbank_dir / f"{process_id}_dna.gb"
+        # Define GenBank paths if needed
+        if save_genbank:
+            protein_gb_path = output_manager.protein_genbank_dir / f"{process_id}.gb"
+            nucleotide_gb_path = output_manager.nucleotide_genbank_dir / f"{process_id}.gb"
 
         # Check if files already exist
         if (sequence_type in ["protein", "both"] and protein_path.exists()) or (
@@ -91,8 +93,8 @@ def process_sample(
 
                 # Download GenBank file if requested
                 if save_genbank:
-                    save_genbank_file(
-                        processor.entrez, original_id, "protein", protein_gb_path
+                    output_manager.save_genbank_file(
+                       processor.entrez, original_id, "protein", protein_gb_path
                     )
 
             except Exception as e:
@@ -123,8 +125,8 @@ def process_sample(
 
                 # Download GenBank file if requested
                 if save_genbank:
-                    save_genbank_file(
-                        processor.entrez, original_id, "nucleotide", nucleotide_gb_path
+                    output_manager.save_genbank_file(
+                       processor.entrez, original_id, "nucleotide", nucleotide_gb_path
                     )
 
             except Exception as e:
@@ -154,7 +156,7 @@ def process_single_taxid(
     save_genbank: bool = False,
 ) -> None:
     try:
-        # Initialize progress counters
+        # Initialise progress counters
         progress_counters = {
             "sequence_counter": 0,
             "max_sequences": max_sequences,
@@ -192,7 +194,7 @@ def process_single_taxid(
                     nucleotide_records = nucleotide_records[:max_sequences]
 
         # Create output manager
-        output_manager = OutputManager(output_dir)
+        output_manager = OutputManager(output_dir, save_genbank)
 
         # Save protein sequences
         if sequence_type in ["protein", "both"] and protein_records:
@@ -202,7 +204,7 @@ def process_single_taxid(
 
                 # Save FASTA
                 filename = f"{record.id}.fasta"
-                output_path = output_dir / filename
+                output_path = output_manager.protein_dir / filename
                 SeqIO.write(record, output_path, "fasta")
                 logger.info(
                     f"Written protein sequence {i+1}/{len(protein_records)} to '{output_path}'"
@@ -210,8 +212,8 @@ def process_single_taxid(
 
                 # Save GenBank if requested
                 if save_genbank:
-                    gb_path = output_manager.genbank_dir / f"{record.id}.gb"
-                    save_genbank_file(processor.entrez, original_id, "protein", gb_path)
+                    gb_path = output_manager.protein_genbank_dir / f"{record.id}.gb"
[…]
