+++ gene_fetch.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #SBATCH --job-name=GF
-#SBATCH --partition=day
+#SBATCH --partition=medium
 #SBATCH --output=%x_%j.out
 #SBATCH --error=%x_%j.err
 #SBATCH --mem=4G
@@ -9,14 +9,15 @@
 
 
 # Activate conda env
-source path/to/profile.d/conda.sh
-conda activate gene-fetch
+#source path/to/profile.d/conda.sh
+#conda activate gene-fetch
 
 
 
 # NCBI API parameters
-EMAIL=user_email@domain.ac.uk
-API_KEY= XYZ
+EMAIL=d.parsons@nhm.ac.uk
+
+API_KEY="1866f9734a06f26bc5895a84387542ac9308"
 
 
 # Gene search variables
@@ -27,16 +28,20 @@ TYPE=both
 
 
 # Input and output directory paths
-SAMPLES_CSV=path/to/input/samples.csv
-OUTPUT_DIR=path/to/out/dir
+SAMPLES_CSV=./data/samples.csv
+#TAXONOMY_CSV=./data/samples_taxonomy.csv
+
+OUTPUT_DIR=./test_samples
 
 
 
 # Run gene_fetch.py
 gene-fetch \
     --gene $GENE \
-    --type $TYPE
+    --type $TYPE \
     --out $OUTPUT_DIR \
     --in $SAMPLES_CSV \
 	--email $EMAIL \
-	--api-key $API_KEY
+	--api-key $API_KEY 
+	#\
+	#-gb
