### 2025-07-07 b9a4246
Update gene_fetch.sh
