### 2025-07-07 abbe0eb
issue #9 and #10 pytest fixes
