+++ tests/test_processors.py
@@ -47,23 +47,20 @@ def mock_output_manager():
     """Create mock OutputManager for testing."""
     with tempfile.TemporaryDirectory() as tmpdirname:
         output_dir = Path(tmpdirname)
-        manager = OutputManager(output_dir)
-        
-        # Create expected directories
-        manager.nucleotide_dir.mkdir(exist_ok=True)
-        manager.genbank_dir.mkdir(exist_ok=True)
+        # Create OutputManager with GenBank enabled for testing
+        manager = OutputManager(output_dir, save_genbank=True)
         
         # Spy on write methods using MagicMock
         manager.write_sequence_reference = MagicMock(wraps=manager.write_sequence_reference)
         manager.log_failure = MagicMock(wraps=manager.log_failure)
         manager.save_sequence_summary = MagicMock(wraps=manager.save_sequence_summary)
+        manager.save_genbank_file = MagicMock(wraps=manager.save_genbank_file)
         
         yield manager
 
 
 @patch('gene_fetch.processors.SeqIO')
-@patch('gene_fetch.processors.save_genbank_file')
-def test_process_sample_both_types(mock_save_genbank, mock_seqio, mock_processor, mock_output_manager):
+def test_process_sample_both_types(mock_seqio, mock_processor, mock_output_manager):
     """Process sample with both protein and nucleotide types."""
     # Set up test parameters
     process_id = "TEST123"
@@ -91,7 +88,7 @@ def test_process_sample_both_types(mock_save_genbank, mock_seqio, mock_processor
     assert mock_seqio.write.call_count == 2  # Once for protein, once for nucleotide
     
     # Check GenBank files were saved
-    assert mock_save_genbank.call_count == 2  # Once for protein, once for nucleotide
+    assert mock_output_manager.save_genbank_file.call_count == 2  # Once for protein, once for nucleotide
     
     # Check sequence reference was written
     mock_output_manager.write_sequence_reference.assert_called_once()
@@ -101,8 +98,7 @@ def test_process_sample_both_types(mock_save_genbank, mock_seqio, mock_processor
 
 
 @patch('gene_fetch.processors.SeqIO')
-@patch('gene_fetch.processors.save_genbank_file')
-def test_process_sample_protein_only(mock_save_genbank, mock_seqio, mock_processor, mock_output_manager):
+def test_process_sample_protein_only(mock_seqio, mock_processor, mock_output_manager):
     """Process sample with protein type only."""
     # Set up test parameters
     process_id = "TEST123"
@@ -130,7 +126,7 @@ def test_process_sample_protein_only(mock_save_genbank, mock_seqio, mock_process
     assert mock_seqio.write.call_count == 1
     
     # Check no GenBank files were saved (save_genbank=False)
-    mock_save_genbank.assert_not_called()
+    mock_output_manager.save_genbank_file.assert_not_called()
     
     # Check sequence reference was written
     mock_output_manager.write_sequence_reference.assert_called_once()
@@ -181,37 +177,43 @@ def test_process_sample_no_sequences(mock_seqio, mock_processor, mock_output_man
 
 
 @patch('gene_fetch.processors.SeqIO')
-@patch('gene_fetch.processors.make_out_dir')
-@patch('gene_fetch.processors.save_genbank_file')
-def test_process_single_taxid(mock_save_genbank, mock_make_dir, mock_seqio, mock_processor, mock_output_manager):
+@patch('gene_fetch.processors.OutputManager')
+def test_process_single_taxid(mock_output_manager_class, mock_seqio, mock_processor):
     """Process single taxid to fetch all available sequences."""
     # Set up test parameters
     taxid = "9606"
     gene_name = "rbcl"
[…]
