+++ tests/test_output_manager.py
@@ -13,20 +13,35 @@ from Bio.Seq import Seq
 
 @pytest.fixture
 def output_manager():
-    """Create OutputManager instance for testing with temp dir."""
+    """Create OutputManager instance for testing with temp dir (no GenBank)."""
     with tempfile.TemporaryDirectory() as tmpdirname:
         # Create OutputManager with the temporary directory
         output_dir = Path(tmpdirname)
-        manager = OutputManager(output_dir)
+        manager = OutputManager(output_dir, save_genbank=False)
         yield manager
 
 
-def test_initialization(output_manager):
-    """OutputManager initialises correctly with required dirs and files."""
+@pytest.fixture
+def output_manager_with_genbank():
+    """Create OutputManager instance for testing with temp dir (with GenBank)."""
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        # Create OutputManager with the temporary directory and GenBank enabled
+        output_dir = Path(tmpdirname)
+        manager = OutputManager(output_dir, save_genbank=True)
+        yield manager
+
+
+def test_initialisation(output_manager):
+    """OutputManager initialises correctly with required dirs and files (no GenBank)."""
     # Check if dirs created
     assert output_manager.output_dir.exists()
+    assert output_manager.protein_dir.exists()
     assert output_manager.nucleotide_dir.exists()
-    assert output_manager.genbank_dir.exists()
+    
+    # GenBank directories should not be created when save_genbank=False
+    assert output_manager.genbank_dir is None
+    assert output_manager.protein_genbank_dir is None
+    assert output_manager.nucleotide_genbank_dir is None
     
     # Check if files created with correct headers
     assert output_manager.failed_searches_path.exists()
@@ -48,6 +63,19 @@ def test_initialization(output_manager):
             "ncbi_taxonomy", "reference_name", "protein_reference_path", 
             "nucleotide_reference_path"
         ]
+
+
+def test_initialisation_with_genbank(output_manager_with_genbank):
+    """OutputManager initialises correctly with GenBank dirs when enabled."""
+    # Check if dirs created
+    assert output_manager_with_genbank.output_dir.exists()
+    assert output_manager_with_genbank.protein_dir.exists()
+    assert output_manager_with_genbank.nucleotide_dir.exists()
+    
+    # GenBank directories should be created when save_genbank=True
+    assert output_manager_with_genbank.genbank_dir.exists()
+    assert output_manager_with_genbank.protein_genbank_dir.exists()
+    assert output_manager_with_genbank.nucleotide_genbank_dir.exists()
 
 
 def test_log_failure(output_manager):
@@ -134,7 +162,7 @@ def test_save_sequence_summary(output_manager):
 
 
 @patch('gene_fetch.output_manager.EntrezHandler')
-def test_save_genbank_file(mock_entrez, output_manager):
+def test_save_genbank_file(mock_entrez, output_manager_with_genbank):
     """Save GenBank file fetched from NCBI."""
     # Create mock handle with test content
     mock_handle = MagicMock()
@@ -147,10 +175,10 @@ def test_save_genbank_file(mock_entrez, output_manager):
     # Define test parameters
     record_id = "NM_12345"
     db = "nucleotide"
-    output_path = output_manager.genbank_dir / "test_genbank.gb"
+    output_path = output_manager_with_genbank.nucleotide_genbank_dir / "test_genbank.gb"
     
     # Call method
-    result = output_manager.save_genbank_file(mock_entrez_instance, record_id, db, output_path)
+    result = output_manager_with_genbank.save_genbank_file(mock_entrez_instance, record_id, db, output_path)
     
     # Verify result and file
     assert result is True
[…]
