### 2025-07-07 bge-barcoding/gene_fetch#10 @SchistoDan
- See [issue 9](https://github.com/bge-barcoding/gene_fetch/issues/9#event-18473463236) regarding the addition of `data/` dir to the repo which contains the 'samples.csv' and 'samples_taxonomy.csv' files mentioned in README.md

- Regarding the missing genus or species - the genus was provided but the species wasn't, although this is not the expected behaviour - thanks for spotting this error. Despite the taxonomic traversal being implemented, the taxonomy csv validation check was happening before the taxonomy traversal, so it was failing unless a genus AND species were provided. I've updated `process_taxonomy_csv` in `processor.py`, making it less restrictive, so that it now just validates that at least some taxonomic information (i.e. at least one rank) was provided. I've rerun the taxonomy.csv test input data after the update and it works as expected.

Unfortunetly, I couldn't replicate the "no valid sequences found for taxID ..." warnings - both nucleotide and protein sequences were found for sample-2 (taxid 30147) and sample-3 (taxid 1875331). What was the example command you used?

Please see commit [b9a4246]https://github.com/SchistoDan/gene_fetch/commit/b9a42468210bbe391d307c8adec3cc3f07ed3293() for changes. The corresponding tests have also been [updated accordingly](https://github.com/SchistoDan/gene_fetch/commit/abbe0eb4ea84257f22e1bab9f8d25aa3525e4708).
