### 2025-07-07 bge-barcoding/gene_fetch#9 @hrhotz
very nice !
Now, I guess, the director structure display in the README.md file just need some adjustments (in order to be in sync with your latest changes.
(and while you are updating the README.md file, you might wanna fix the arguments list (e.g.: adding "-t"
