### 2025-07-07 bge-barcoding/gene_fetch#9 @SchistoDan
Ah yes, thank you! I'll update the README.md now
