### 2025-07-07 bge-barcoding/gene_fetch#10 @SchistoDan
See commit [97778f3](https://github.com/SchistoDan/gene_fetch/commit/97778f3251c103ff7bb29a1fef816f33138ae40b#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5) for the README.md update mentioned in my previous comment.
