### 2025-07-07 1a8ea33
updates to checkpointing logic, --clean, README (issue #11)
