+++ standalone/gene_fetch.py
@@ -1,4020 +1,0 @@
-"""
-Gene Fetch - NCBI Sequence Retrieval Tool
-
-This script fetches sequence data from NCBI databases
-using sample taxonomic information.
-
-Input:
-- NCBI account email address and API key (see:
-  https://support.nlm.nih.gov/kbArticle/?pn=KA-05317)
-- CSV file containing sample IDs, and taxonomy IDs or taxonomic heirarchy
-- Gene name (e.g., 'cox1', '16s', 'rbcl', 'matk')
-- Output directory path (will create new directories)
-- Sequence type ('protein', 'nucleotide', or 'both')
-- Optional: Minimum nucleotide/protein sequence length thresholds
-- Optional: 'Single' taxid mode (or default 'batch' taxid mode)
-- Optional: Maximum number of sequences to fetch ('single' taxid mode only)
-- Optional: Retrieve corresponding genbank entries for each sequence
-
-Output:
-- FASTA files containing retrieved sequences (named by ID)
-- Log file
-- CSV file containing fetched sequence metadata
-- CSV file documenting failed retrievals with error types
-- 'Single' mode: CSV file(s) containing accession number, length, and sequence
-
-Dependencies:
-- Python>=3.9
-- Biopython>=1.80
-- ratelimit>=2.2.1
-
-Usage:
-    python gene_fetch.py -g/--gene <gene_name> -o/--out <output_directory>
-                    --type <sequence_type> -i/--in <samples.csv>
-                    [-i1/--in2 <sameples_taxonomy.csv] [-s/--single <taxid>]
-                    [--protein-size <min_size>] [--nucleotide-size <min_size>]
-                    [-e/--email <email>] [-k/--api-key <api_key>]
-                    [--max-sequences <max_num>]
-
-    # Required arguments:
-    -e/--email           Email to use for NCBI API requests
-    -k/--api-key         API key to use for NCBI API requests
-    -g/--gene            Gene name to search for (e.g., cox1, 16s, rbcl)
-    -o/--out             Directory to save output files
-    --type               Sequence type to fetch (protein, nucleotide, or both)
-    -i/--in              Input CSV file with taxonomy IDs (required unless
-                         using --single)
-    -i2/--in2            Input CSV file with taxonomic heirarchy (required
-                         unless using --single)
-
-    # Optional arguments:
-    -s/--single            Single taxID to fetch all available sequences for
-    --protein-size         Minimum protein sequence length (default: 500)
-    --nucleotide-size      Minimum nucleotide sequence length (default: 1000)
-    --max-sequences        Maximum number of sequences to retrieve (in
-                           single mode) (default: all)
-
-
-Future Development:
-- Add optional alignment of retrieved sequences
-- Add support for additional genetic markers beyond the currently supported set
-
-Author: D. Parsons & B. Price @ NHMUK
-Version: 1.0
-License: MIT
-"""
-
-# =============================================================================
-# Imports and setup
-# =============================================================================
-import csv
-import sys
-import time
-import argparse
-import logging
-from Bio import Entrez, SeqIO
-from functools import wraps
-from ratelimit import limits, sleep_and_retry
-from time import sleep
-from random import uniform
-from urllib.error import HTTPError
-from dataclasses import dataclass
-from typing import Optional, Tuple, List, Dict, Any
-from Bio.SeqRecord import SeqRecord
-from pathlib import Path
-from Bio.Seq import Seq
-from http.client import IncompleteRead
-import re
-
-
-def setup_argument_parser():
-    parser = argparse.ArgumentParser(
-        description="Fetch gene sequences from NCBI databases."
-    )
-
-    parser.add_argument(
-        "-g",
-        "--gene",
-        required=True,
[…]
