### 2025-07-08 #8456 @hrhotz
@SchistoDan and I have ticked all the boxes in the review form
Once again, it has been a very enjoyable experience doing a review for JOSS - Thank you @abhishektiwari
