### 2025-07-10 bge-barcoding/gene_fetch#14 @ManavalanG
Unit tests failing
Regarding https://github.com/openjournals/joss-reviews/issues/8456#issuecomment-3048196421:

Running pytest fails for some unit tests. Here is the summary section:

```
================================================================================ short test summary info ================================================================================
FAILED tests/test_main.py::test_main_no_input_files - AssertionError: assert 1 == 0
FAILED tests/test_processors.py::test_process_single_taxid - AssertionError: expected call not found.
ERROR tests/test_output_manager.py::test_initialisation - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_initialisation_with_genbank - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_log_failure - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_write_sequence_reference - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_save_sequence_summary - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_save_genbank_file - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_save_genbank_file_error - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_save_genbank_file_disabled - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_initialisation_no_sequence_refs - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_output_manager.py::test_write_sequence_reference_disabled - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_sample_both_types - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_sample_protein_only - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_sample_no_sequences - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_taxid_csv - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_taxonomy_csv - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_taxonomy_csv_missing_columns - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
ERROR tests/test_processors.py::test_process_taxonomy_csv_taxid_not_found - TypeError: __init__() got an unexpected keyword argument 'save_genbank'
====================================================== 2 failed, 53 passed, 8 deselected, 1 warning, 17 errors in 60.19s (0:01:00) ======================================================
```

I'm on mac and using the `main` branch. Conda environment I used:

```yaml
name: gene-fetch
channels:
  - conda-forge
  - bioconda
dependencies:
  - biopython=1.85
  - gene-fetch=1.0.12
  - ratelimit=2.2.1
  - python=3.9
  - pytest
```

Same result with python versions 3.11 and 3.13.

<summary>Here is the full stderr/stdout:</summary>
$ pytest
================================================================================== test session starts ==================================================================================
platform darwin -- Python 3.9.23, pytest-8.4.1, pluggy-1.6.0
rootdir: /Documents/other_repos/gene_fetch
configfile: pytest.ini
testpaths: tests
collected 80 items / 8 deselected / 72 selected

tests/test_core.py ...............                                                                                                                                                [ 20%]
tests/test_entrez_handler.py .........                                                                                                                                            [ 33%]
tests/test_integration.py ..                                                                                                                                                      [ 36%]
tests/test_main.py ......F..                                                                                                                                                      [ 48%]
tests/test_output_manager.py EEEEEEEEEE                                                                                                                                           [ 62%]
tests/test_processors.py EEEF.EEEE                                                                                                                                                [ 75%]
tests/test_sequence_processor.py .................                                                                                                                                [ 98%]
tests/test_simple_import.py .                                                                                                                                                     [100%]

======================================================================================== ERRORS =========================================================================================
_________________________________________________________________________ ERROR at setup of test_initialisation _________________________________________________________________________

    @pytest.fixture
    def output_manager():
        """Create OutputManager instance for testing with temp dir (no GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:20: TypeError
__________________________________________________________________ ERROR at setup of test_initialisation_with_genbank ___________________________________________________________________

    @pytest.fixture
    def output_manager_with_genbank():
        """Create OutputManager instance for testing with temp dir (with GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory and GenBank enabled
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:30: TypeError
__________________________________________________________________________ ERROR at setup of test_log_failure ___________________________________________________________________________

    @pytest.fixture
    def output_manager():
        """Create OutputManager instance for testing with temp dir (no GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:20: TypeError
____________________________________________________________________ ERROR at setup of test_write_sequence_reference ____________________________________________________________________

    @pytest.fixture
    def output_manager():
        """Create OutputManager instance for testing with temp dir (no GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:20: TypeError
_____________________________________________________________________ ERROR at setup of test_save_sequence_summary ______________________________________________________________________

    @pytest.fixture
    def output_manager():
        """Create OutputManager instance for testing with temp dir (no GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:20: TypeError
_______________________________________________________________________ ERROR at setup of test_save_genbank_file ________________________________________________________________________

    @pytest.fixture
    def output_manager_with_genbank():
        """Create OutputManager instance for testing with temp dir (with GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory and GenBank enabled
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:30: TypeError
____________________________________________________________________ ERROR at setup of test_save_genbank_file_error _____________________________________________________________________

    @pytest.fixture
    def output_manager_with_genbank():
        """Create OutputManager instance for testing with temp dir (with GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory and GenBank enabled
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:30: TypeError
___________________________________________________________________ ERROR at setup of test_save_genbank_file_disabled ___________________________________________________________________

    @pytest.fixture
    def output_manager():
        """Create OutputManager instance for testing with temp dir (no GenBank)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:20: TypeError
________________________________________________________________ ERROR at setup of test_initialisation_no_sequence_refs _________________________________________________________________

    @pytest.fixture
    def output_manager_no_refs():
        """Create OutputManager instance for testing with temp dir (no sequence refs)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory but no sequence references
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=False)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:241: TypeError
_______________________________________________________________ ERROR at setup of test_write_sequence_reference_disabled ________________________________________________________________

    @pytest.fixture
    def output_manager_no_refs():
        """Create OutputManager instance for testing with temp dir (no sequence refs)."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            # Create OutputManager with the temporary directory but no sequence references
            output_dir = Path(tmpdirname)
>           manager = OutputManager(output_dir, save_genbank=False, create_sequence_refs=False)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_output_manager.py:241: TypeError
___________________________________________________________________ ERROR at setup of test_process_sample_both_types ____________________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
__________________________________________________________________ ERROR at setup of test_process_sample_protein_only ___________________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
__________________________________________________________________ ERROR at setup of test_process_sample_no_sequences ___________________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
_______________________________________________________________________ ERROR at setup of test_process_taxid_csv ________________________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
______________________________________________________________________ ERROR at setup of test_process_taxonomy_csv ______________________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
______________________________________________________________ ERROR at setup of test_process_taxonomy_csv_missing_columns ______________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
______________________________________________________________ ERROR at setup of test_process_taxonomy_csv_taxid_not_found ______________________________________________________________

    @pytest.fixture
    def mock_output_manager():
        """Create mock OutputManager for testing."""
        with tempfile.TemporaryDirectory() as tmpdirname:
            output_dir = Path(tmpdirname)
            # Create OutputManager with GenBank enabled and sequence refs enabled for testing
>           manager = OutputManager(output_dir, save_genbank=True, create_sequence_refs=True)
E           TypeError: __init__() got an unexpected keyword argument 'save_genbank'

tests/test_processors.py:51: TypeError
======================================================================================= FAILURES ========================================================================================
_______________________________________________________________________________ test_main_no_input_files ________________________________________________________________________________

mock_parser = <MagicMock name='setup_argument_parser' id='4412630832'>, mock_save_run_info = <MagicMock name='save_run_info' id='4417820176'>
mock_clear_output_dir = <MagicMock name='clear_output_directory' id='4417840464'>, mock_should_clear_output_dir = <MagicMock name='should_clear_output_directory' id='4417852560'>
mock_make_out_dir = <MagicMock name='make_out_dir' id='4417868752'>, mock_setup_logging = <MagicMock name='setup_logging' id='4417880848'>
mock_config = <MagicMock name='Config' id='4417892944'>, mock_entrez = <MagicMock name='EntrezHandler' id='4417909136'>
mock_processor = <MagicMock name='SequenceProcessor' id='4417921232'>, mock_exit = <MagicMock name='exit' id='4417933264'>

    @patch('gene_fetch.main.sys.exit')
    @patch('gene_fetch.main.SequenceProcessor')
    @patch('gene_fetch.main.EntrezHandler')
    @patch('gene_fetch.main.Config')
    @patch('gene_fetch.main.setup_logging')
    @patch('gene_fetch.main.make_out_dir')
    @patch('gene_fetch.main.should_clear_output_directory', create=True)
    @patch('gene_fetch.main.clear_output_directory', create=True)
    @patch('gene_fetch.main.save_run_info', create=True)
    @patch('gene_fetch.main.setup_argument_parser')
    def test_main_no_input_files(mock_parser, mock_save_run_info, mock_clear_output_dir, mock_should_clear_output_dir,
                                mock_make_out_dir, mock_setup_logging,
                                mock_config, mock_entrez, mock_processor,
                                mock_exit):
        """Test that main function completes when argparse validation is bypassed via mocking."""
        # Create a mock parser that returns a namespace with no input files
        mock_parser_instance = MagicMock()
        mock_args = argparse.Namespace(
            gene='cox1',
            out='/tmp/out',
            type='both',
            email='test@example.com',
            api_key='valid_test_key_12345',
            input_csv=None,
[…]
