### 2025-07-14 bge-barcoding/gene_fetch#14 @SchistoDan
Thanks @ManavalanG for sending of the full error - I am, however, having trouble reproducing them.

I see you're installing v1.0.12 of gene-fetch. Could you please update to v1.0.13, either via [pypi](https://pypi.org/project/gene-fetch/) or [bioconda](https://bioconda.github.io/recipes/gene-fetch/README.html#package-package%20&#x27;gene-fetch&#x27;), or a dev installation with `pip install -e .` when in the cloned gene_fetch dir.

If you're still getting errors, please let me know.
