### 2025-07-22 bge-barcoding/gene_fetch#14 @ManavalanG
Upgrading to `v1.0.13` resolved the issue. Thanks! Please update the gene-fetch version in `environment.yaml` to reflect this.
