### 2025-07-22 745d91d
Update environment.yaml
Update gene-fetch dependency to 1.0.13
