+++ environment.yaml
@@ -4,5 +4,5 @@ channels:
   - bioconda
 dependencies:
   - biopython=1.85
-  - gene-fetch=1.0.12
+  - gene-fetch=1.0.13
   - ratelimit=2.2.1
