### 2025-07-24 bge-barcoding/gene_fetch#15 @ManavalanG
Version info option is missing [JOSS]
Regarding https://github.com/openjournals/joss-reviews/issues/8456#issuecomment-3048196421:

`gene-fetch` currently does not seem to offer any way to easily check its version. Providing this info either via `--version` or as part of its help message would be helpful.
