### 2025-07-24 bge-barcoding/gene_fetch#16 @ManavalanG
Relatively minor improvements [JOSS]
Regarding https://github.com/openjournals/joss-reviews/issues/8456#issuecomment-3048196421:

This issue will host a live checklist of minor improvements needed as part of the JOSS review. I will update the list as I come across them.

- [x] In the second example command, there is a typo: `-i2 .data/samples_taxonomy.csv ` needs to be `-i2 ./data/samples_taxonomy.csv`.
- [x] Clarify in the opening description in readme that gene fetch deals with NCBI GenBank database - `Gene Fetch enables high-throughput retreival of sequence data from NCBI databases`
- [x] In the testing section of readme, adding a note that `pip install -e .` helps with tool installation from the source code would be helpful before running the pytest. This would allow the user to test the source code in development.
- [x] It would benefit the user to clarify how the sequence matches are made. That is, they are retrieved based on the sequence annotation available in Genbank and not based on any type of sequence similarity checking.
- [x] Manuscript uses the term `genetic markers` but the readme doc calls them as `targets`. Use of same terms in both would help avoid confusion.
- [x] Both manuscript and readme lists support for 16 genetic markets/targets. `src/gene_fetch/core.py` additionally includes support for `psbA` and `23s`, but they appear to be excluded from the claims. Why? Are they currently experimental?
- [x] `Supported targets` section in readme says `More targets can be added if necessary (see 'class config')`; please expand on this to help the user customize for their own genetic marker. Mention the file path that contains `class config`.
- [x] [Nice to have but not needed to make this change] Currently hard to tell which commit corresponds to which version. I would recommend github releases with version info, git tagging or both. Again, this is a nice-to-have but not needed.
