### 2025-07-24 bge-barcoding/gene_fetch#17 @ManavalanG
Some retrieved sequences don't match the query gene [JOSS]
Regarding https://github.com/openjournals/joss-reviews/issues/8456#issuecomment-3048196421:

The third example code from readme:

```sh
gene-fetch -e $EMAIL -k $API \
            -g rbcL -o ./output/out3 -s 3702 \
            --type protein --protein-size 400 --max-sequences 1000
```

returns 30 hits. While 28 of them are good match to the query gene `rbcL`, two of them are not (`MRL1` and `PDE338`):

```
Accession,Length,Description
Q9M9H4.1,500,"RecName: Full=Protein PIGMENT DEFECTIVE 338, chloroplastic; AltName: Full=Protein PETB/PETD STABILIZING FACTOR; AltName: Full=Protein PHOTOSYSTEM BIOGENESIS REGULATOR 1; AltName: Full=Protein rbcL RNA S1-binding domain; Flags: Precursor"
Q0WLC6.2,1089,"RecName: Full=Pentatricopeptide repeat-containing protein MRL1, chloroplastic; AltName: Full=Protein MATURATION OF RBCL 1; Short=AtMRL1; Flags: Precursor"
```

From the query string constructed by gene-fetch:

```
2025-07-23 22:31:27,298 [INFO] Searching protein database at rank species (Arabidopsis thaliana) with term: (rbcL[Gene] OR RBCL[Gene] OR "ribulose-1,5-bisphosphate carboxylase/oxygenase large subunit"[Protein Name] OR "ribulose 1,5-bisphosphate carboxylase/oxygenase large subunit"[Protein Name] OR "ribulose bisphosphate carboxylase large chain"[Protein Name] OR "RuBisCO large subunit"[Protein Name] OR "ribulose-1,5-bisphosphate carboxylase/oxygenase small subunit"[Protein Name] OR "ribulose 1,5-bisphosphate carboxylase/oxygenase small subunit"[Protein Name] OR "ribulose-1,5-bisphosphate carboxylase/oxygenase small chain"[Protein Name] OR "RuBisCO small subunit"[Protein Name] OR "Ribulose bisphosphate carboxylase/oxygenase activase"[Gene] OR "rbcL gene"[Gene] OR "RBCL gene"[Gene] OR rbcL[Title] OR RBCL[Title]) AND txid3702[Organism:exp] AND 400:10000[SLEN]
```

it appears that the above-mentioned two hits are substring matches resulting from `...OR rbcL[Title] OR RBCL[Title]...`. Is this an expected outcome? I don't remember either the manuscript or the readme mentioning that gene-fetch extracts sequence for the exact query gene; however if this is an expected outcome, a note about such non-target retrieval would be helpful to the user.
