### 2025-07-24 bge-barcoding/gene_fetch#18 @ManavalanG
Output directory structure does not match the documented structure in readme [JOSS]
Regarding https://github.com/openjournals/joss-reviews/issues/8456#issuecomment-3048196421:

Output directory structure for the example commands 1 and 2 in batch mode does not match the [documented dir structure](https://github.com/bge-barcoding/gene_fetch?tab=readme-ov-file#batch-mode). I ran the most recent version of gene-fetch in macOS.

For example cmd 1:
```
out1
├── failed_searches.csv
├── genbank
│   ├── failed_searches.csv
│   ├── genbank
│   ├── nucleotide
│   ├── sample-1_dna.gb
│   ├── sample-1.gb
│   ├── sample-2_dna.gb
│   ├── sample-2.gb
│   ├── sample-3_dna.gb
│   ├── sample-3.gb
│   ├── sample-4_dna.gb
│   ├── sample-4.gb
│   └── sequence_references.csv
├── gene_fetch.log
├── nucleotide
│   ├── sample-1_dna.fasta
│   ├── sample-2_dna.fasta
│   ├── sample-3_dna.fasta
│   └── sample-4_dna.fasta
├── sample-1.fasta
├── sample-2.fasta
├── sample-3.fasta
├── sample-4.fasta
└── sequence_references.csv
```

For example cmd 2:

```
out2
├── failed_searches.csv
├── genbank
├── gene_fetch.log
├── nucleotide
│   ├── sample-2_dna.fasta
│   └── sample-3_dna.fasta
└── sequence_references.csv
```
