### 2025-07-28 032e519
--version & typo fix [JOSS issue 15]
