### 2025-07-28 bge-barcoding/gene_fetch#15 @SchistoDan
Thanks @ManavalanG - Great idea.

```
gene-fetch --version
````
Now prints the current version, e.g.:
```
gene-fetch 1.0.13
```
It also prints upon start up:
```
======   Starting Gene Fetch   ======
Version: 1.0.13
Written by Dan Parsons & Ben Price, Natural History Museum London
```

See commit [9c9ef3e](https://github.com/bge-barcoding/gene_fetch/commit/9c9ef3ea21c62bb63c28ddf3ca7fb9f246e362be).
