### 2025-07-28 bge-barcoding/gene_fetch#16 @SchistoDan
@ManavalanG - Thank you for raising these valuable improvements to the readme.

1. Example command corrected as highlighted.
2. Opening description text changed to "NCBI's GenBank sequence database" to improve accuracy of statement.
3. Added relevant optional 'step' in testing section (line 94 & 95).

See commit [dc714d1](https://github.com/bge-barcoding/gene_fetch/commit/dc714d1cacb59cd28a991245c6551a8e474cf51e)
