### 2025-07-28 a704d15
rbcL title search correction [JOSS issue 17]
