+++ src/gene_fetch/__init__.py
@@ -5,7 +5,7 @@ This package fetches sequence data from NCBI databases
 using sample taxonomic information.
 """
 
-__version__ = "1.0.13"
+__version__ = "1.0.14"
 
 from .core import Config
 from .entrez_handler import EntrezHandler
