### 2025-07-28 bge-barcoding/gene_fetch#17 @SchistoDan
This is was not intended - `[Title]` searches were implemented during initial Gene Fetch development but were removed from most other search terms as a title/descption search often returned unintended matches (as you've found). It appears `[Title]` searches were not sufficiently removed from the rbcL search term constructor. I have now removed the `[Title]` searches and run the 3rd example command, which now correctly returns 28 proper hits.

Thanks for raising awreness of this issue @ManavalanG - I have updated the [PyPi package to v1.0.14](https://pypi.org/project/gene-fetch/) to reflect this change in search logic.
