### 2025-07-28 bge-barcoding/gene_fetch#18 @SchistoDan
Hi @ManavalanG - I am having trouble reproducing the output directory structure you mention for both commands. Running the 1st and second example commands with both version 1.0.13 and 1.0.14 have produced the following output directory structures:

Example 1:
```
output_dir
├── .gene_fetch_run_info
├── failed_searches.csv
├── gene_fetch.log
├── sequence_references.csv
├── genbank/
│   ├── nucleotide/
│   │   ├── sample-1.gb
│   │   ├── sample-2.gb
│   │   └── sample-3.gb
│   └── protein/
│       ├── sample-1.gb
│       ├── sample-2.gb
│       └── sample-3.gb
├── nucleotide/
│   ├── sample-1.fasta
│   ├── sample-2.fasta
│   └── sample-3.fasta
└── protein/
    ├── sample-1.fasta
    ├── sample-2.fasta
    └── sample-3.fasta
```
Example 2:
```
output_dir2
├── .gene_fetch_run_info
├── failed_searches.csv
├── gene_fetch.log
├── sequence_references.csv
├── nucleotide/
│   ├── sample-1.fasta
│   ├── sample-2.fasta
│   └── sample-3.fasta
└── protein/
```

If you wouldn't mind updating to the most recent version (1.0.14) and running the two example commands again to see if that solves the issue, that would be greatly appreciated. I previously overlooked updating `environment.yaml` to v1.0.13 (apologies), so you may have been working with 1.0.12 - I have now update `environment.yaml` to list the most up-to-date version, 1.0.14
