### 2025-08-05 bge-barcoding/gene_fetch#17 @ManavalanG
New version checks out for the example I mentioned. However, `src/gene_fetch/core.py` still uses `[Title]` for other potential query genes. Wouldn't they be affected by the same issue as `rbcL`?
