### 2025-08-05 cd29921
JOSS issues 16,17,19
