+++ src/gene_fetch/core.py
@@ -262,40 +262,55 @@ class Config:
         self.min_protein_size_single_mode = 100
 
         self.gene_search_term = ""
+        
+        # Define target aliases for common gene name variations
+        self._target_aliases = {
+            "coi": "cox1",
+            "co1": "cox1",
+            "cox1": "cox1",
+            "coii": "cox2",
+            "co2": "cox2",
+            "cox2": "cox2",
+            "coiii": "cox3",
+            "co3": "cox3",
+            "cox3": "cox3",
+            "cytb": "cytb",
+            "cob": "cytb",
+            "nd1": "nd1",
+            "nad1": "nd1",
+            "nd2": "nd2",
+            "nad2": "nd2",
+            "rbcl": "rbcl",
+            "rbcL": "rbcl",
+            "matk": "matk",
+            "matK": "matk",
+            "psba": "psba",
+            "psbA": "psba",
+            "trnh-psba": "psba",
+            "psba-trnh": "psba",
+        }
     
         # Define gene type categories
-        self._rRNA_genes = {
+        self._rRNA_genes = {              
             "16s": [
                 "16S ribosomal RNA[Gene]",
                 "16S rRNA[Gene]",
                 "rrs[Gene]",
                 "rrn16[Gene]",
-                "16S ribosomal RNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "16S rRNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "16S[Title] NOT methylase[Title] NOT methyltransferase[Title] "
-                "NOT pseudouridylate[Title] NOT synthase[Title]",
-                "16S ribosomal RNA[All Fields]",
-                "rrs[All Fields]",
-                "rrn16[All Fields]",
+                "16s[Gene]",
+                "16s[rRNA]",
+                "16S ribosomal RNA[rRNA]",
+                "16S rRNA[rRNA]",
+                "rrs[rRNA]",
+                "rrn16[rRNA]",
             ],
             "18s": [
                 "18S ribosomal RNA[Gene]",
                 "18S rRNA[Gene]",
                 "rrn18[Gene]",
                 "SSU rRNA[Gene]",
-                "18S ribosomal RNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "18S rRNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "18S[Title] NOT methylase[Title] NOT methyltransferase[Title] "
-                "NOT pseudouridylate[Title] NOT synthase[Title]",
                 "18S ribosomal RNA[rRNA]",
+                "18S rRNA[rRNA]",
                 "SSU rRNA[rRNA]",
                 "rrn18[rRNA]",
             ],
@@ -304,15 +319,8 @@ class Config:
                 "23S rRNA[Gene]",
                 "rrl[Gene]",
                 "rrn23[Gene]",
-                "23S ribosomal RNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "23S rRNA[Title] NOT methylase[Title] NOT "
-                "methyltransferase[Title] NOT pseudouridylate[Title] NOT "
-                "synthase[Title]",
-                "23S[Title] NOT methylase[Title] NOT methyltransferase[Title] "
[…]

[…]
