### 2025-08-05 bge-barcoding/gene_fetch#16 @SchistoDan
@ManavalanG - Thanks for your additional corrections.

4. See addition to the `highlgihted features` section (line 19 of README) clarifying that sequence matches are made by searching the annotation/feature table of GenBank records.
5. Changed 'genetic marker(s)' to 'target' in the manuscript for consistency.
6. This was an oversight as these targets were added slightly later than the others. The README has been updated to list 18 common targets, including psbA and 23s.
7. The guidance for users to add extra target search terms into `core.py` has been expanded within the `supported targets` section. It now signposts the two relevent dictionaries within the config class of `core.py` where users can add aditional targets. It also references the `Contributions and guidlines` section if users wish to make an Issue/Pull request to get their additional targets included in the main Gene Fetch release.
8. Thanks for the recommendation. After I have addressed all your comments and you have approved them, I will make a tag and corresponding release (and updates to the PyPi and bioconda packages accordingly).

See commit [cd29921](https://github.com/bge-barcoding/gene_fetch/commit/cd29921b03b28116a521c98d4f0c417fdec52399) for changes 4-7.
