### 2025-08-08 #8456 @abhishektiwari
@SchistoDan

**Corrections**
1. Line 3: Second author name Benjamin W. Price does not match with Zenodo archive (Price, Benjamin)
2. Line 4: Incorrect RoR for Natural History Museum (should be 039zvsn29 here is link https://ror.org/039zvsn29)
3. Line 26: "taxa" (plural) instead of "taxon" (singular)??
4. Line 84: "PyPi" should be "PyPI"
5. Lined 110: “von Ammon, U.” not “Ammon, U. von”

**Feedback on summary**

I think we can improve the summary section.

**Firstly**, I suggest moving following to “Statement of need “ or remove it completely. It will shorten the summary section for the reader.

> “GenBank is one of three mirrored partner databases which, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. GenBank currently holds over 3.6 billion nucleotide sequences from over 500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. ”

**Secondly**, summary should describe the high-level functionality and purpose of the software for a diverse, non-specialist audience. Can you check if the remaining summary text can be up-levelled for a diverse, non-specialist audience?
