### 2025-08-08 #8456 @abhishektiwari
@SchistoDan When editing Zenodo you should be able to edit existing version of Zenodo archive without creating a new version.
