### 2025-08-08 0788782
summary & SoN section update
