+++ paper/paper.bib
@@ -14,22 +14,6 @@
     eprint = {https://academic.oup.com/nar/article-pdf/41/D1/D36/3680750/gks1195.pdf},
 }
 
-@article{OCathail2024,
-    author = {O’Cathail, Colman and Ahamed, Alisha and Burgin, Josephine and Cummins, Carla and Devaraj, Rajkumar and Gueye, Khadim and Gupta, Dipayan and Gupta, Vikas and Haseeb, Muhammad and Ihsan, Maira and Ivanov, Eugene and Jayathilaka, Suran and Kadhirvelu, Vishnukumar and Kumar, Manish and Lathi, Ankur and Leinonen, Rasko and McKinnon, Jasmine and Meszaros, Lili and Pauperio, Joana and Pesant, Stephane and Rahman, Nadim and Rinck, Gabriele and Selvakumar, Sandeep and Suman, Swati and Sunthornyotin, Yanisa and Ventouratou, Marianna and Waheed, Zahra and Woollard, Peter and Yuan, David and Zyoud, Ahmad and Burdett, Tony and Cochrane, Guy},
-    title = {The European Nucleotide Archive in 2024},
-    journal = {Nucleic Acids Research},
-    volume = {53},
-    number = {D1},
-    pages = {D49-D55},
-    year = {2024},
-    month = {11},
-    abstract = {The European Nucleotide Archive (ENA, https://www.ebi.ac.uk/ena), maintained at the European Molecular Biology Laboratory's European Bioinformatics Institute (EMBL-EBI) provides freely accessible services, both for deposition of, and access to, open nucleotide sequencing data. Open scientific data are of paramount importance to the scientific community and contribute daily to the acceleration of scientific advance. Outlined here are changes to and updates on the ENA service in 2024, aligning with the broad goals of enhancing interoperability, globalisation of the service and scaling the platform to meet current and future needs.},
-    issn = {1362-4962},
-    doi = {10.1093/nar/gkae975},
-    url = {https://doi.org/10.1093/nar/gkae975},
-    eprint = {https://academic.oup.com/nar/article-pdf/53/D1/D49/60743838/gkae975.pdf},
-}
-
 @article{Oleary2024,
     title = {Exploring and retrieving sequence and metadata for species across the tree of life with NCBI Datasets},
     author = {O'Leary, Nuala A and Cox, Eric and Holmes, J Bradley and Anderson, W Ray and Falk, Robert and Hem, Vichet and Tsuchiya, Mirian TN and Schuler, Gregory D and Zhang, Xuan and Torcivia, John and Ketter, Anne and Breen, Laurie and Cothran, Jonathan and Bajwa, Hena and Tinne, Jovany and Meric, Peter A and Hlavina, Wratko and Schneider, Valerie A},
@@ -43,54 +27,6 @@
     publisher = {Nature Publishing Group},
     doi = {10.1038/s41597-024-03571-y},
     pmid = {38969627}
-}
-
-@article{Fukuda:2020,
-    author = {Fukuda, Asami and Kodama, Yuichi and Mashima, Jun and Fujisawa, Takatomo and Ogasawara, Osamu},
-    title = {DDBJ update: streamlining submission and access of human data},
-    journal = {Nucleic Acids Research},
-    volume = {49},
-    number = {D1},
-    pages = {D71-D75},
-    year = {2020},
-    month = {11},
[…]
