+++ paper/paper.md
@@ -25,27 +25,22 @@ bibliography: paper.bib
 
 # Summary
 
-`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) sequence databases, namely GenBank [@Benson:2012; @Sayers:2024]. GenBank is one of three mirrored partner databases which, along with the European Nucleotide Archive (ENA) [@OCathail2024] and DNA DataBank of Japan (DDBJ) [@Fukuda:2020], form the International Nucleotide Sequence Database Collaboration (INSDC) [@Karsch-Mizrachi:2024]. GenBank currently holds over 3.6 billion nucleotide sequences from over 500,000 formally described species, and exhibits biennial doubling [@Sayers:2023]. 
-
-By streamlining the often tedious and error-prone task of sequence acquisition at scale, `Gene Fetch` addresses a critical need in biological research. It is capable of retrieving both protein and nucleotide sequences for a user-specified target from across the tree of life, including protein-coding genes (such as cytochrome oxidase subunits, NADH dehydrogenases, RuBisCO, and matK), ribosomal RNA genes (like 16S, 18S, and 28S) and the internal transcribed spacer (ITS) regions. The tool accepts both NCBI taxonomic identifiers (taxids) and hierarchical taxonomic information as input.
-
-It can systematically traverse taxonomic hierarchies when target sequences are unavailable at the initially specified taxonomic rank (e.g., species → genus → family, etc), documenting the matched rank. This is especially valuable for researchers working with non-model organisms or taxonomic groups with limited sequence data, facilitating retrieval of the taxonomically closest available sequence. The integrated 'batch' mode processes multiple input taxa and retrieve the single ‘best’ sequence per taxa, whilst 'single' mode exhaustively searches for all target sequences for a specified taxon. Collectively, these modes enable efficient retrieval of sequence data for genomic and phylogenetic studies across diverse taxa.
-
+`Gene Fetch` is an open-source Python tool that automates the retrieval of sequence data from the National Center for Biotechnology Information (NCBI) GenBank sequence database [@Benson:2012; @Sayers:2024]. It simplifies sequence acquisition for biological research by retrieving both protein and nucleotide sequences for user-specified targets from across the tree of life, accepting NCBI taxonomic identifiers or hierarchical taxonomic information as input.
 
 
 
 # Statement of need
 
-Comparative genetic analyses within or between taxonomic groups often requires researchers to gather large numbers of gene or protein sequences from public repositories. Streamlined access to these invaluable databases is essential for comprehensive sequence data analysis, however, this task presents several significant challenges:
-(1) Limited database curation leading to inconsistent sequence annotation.
-(2) Variable sequence representation across taxonomic groups on databases.
+Comparative genetic analyses within or between taxonomic groups often requires researchers to gather large numbers of gene or protein sequences from public repositories such as GenBank, which currently holds over 4.7 billion nucleotide sequences from over 580,000 formally described species [@Sayers:2024]. Streamlined access to this invaluable database is essential for comprehensive sequence data analysis, however, this task presents several significant challenges:
+(1) Limited database curation leading to inconsistent sequence annotations.
+(2) Variable sequence representation across taxonomic groups on the databases.
 (3) Time-intensive manual retrieval processes that do not scale efficiently.
 
[…]
