+++ paper/paper.bib
@@ -57,7 +57,7 @@
 }
 
 @article{Jeunen:2023,
-	author = {Jeunen, Gert-Jan and Dowle, Eddy and Edgecombe, Jonika and Ammon, Ulla {von} and Gemmell, Neil J. and Cross, Hugh},
+	author = {Jeunen, Gert-Jan and Dowle, Eddy and Edgecombe, Jonika and {von Ammon}, Ulla and Gemmell, Neil J. and Cross, Hugh},
 	title = {crabs—A software program to generate curated reference databases for metabarcoding sequencing data},
 	journal = {Molecular Ecology Resources},
 	volume = {23},
