### 2025-08-11 #8456 @abhishektiwari
@SchistoDan - one more correction probably last one

Line 75 in paper.md: `bioconda` should be `Bioconda`.
Line 61 in paper.bib: `crabs` should be `CRABS`
