### 2025-08-11 5193c0d
Update paper.md
Corrected bioconda to Bioconda
