### 2025-08-11 #8456 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
