### 2025-08-14 #8456 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
