### 2025-08-14 #8456 @SchistoDan
Many thanks to all involved for an enjoyable submission and review process 🎉🎉🎉 I look forward to publishing with JOSS again in the future!
